+++ lib/canvas-app/venue/facebook.rb
@@ -48,7 +48,9 @@ module Canvas::App
             qs_oauth = context.connection.auth.venue_token(context.token, 'facebook', venue_data)
             @qs_uuid = context.connection.auth.token_owner(qs_oauth)['uuid']
 
-            context.connection.playercenter.register_player(qs_uuid, game.uuid, 'facebook', context.token)
+            Thread.new do
+              context.connection.playercenter.register_player(qs_uuid, game.uuid, 'facebook', context.token)
+            end
           end
 
           context.tokens = {