+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.0.64 / Unreleased
+
+* Adds invite / login fake avatars
+* Creates venue identities on spiral-galaxy and embedded game plays
+* Moves player registration on game plays into fire and forget threads
+
 # 0.0.63 / 2013-04-07
 
 * Updates SDK app testing dependency